a2a_financial_agent/check_budget_agent/pyproject.toml
[project]
name = "check-budget-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.8",
    "google-adk>=1.3.0",
    "python-dotenv",
    "uvicorn",
    "litellm",
]